Add two new scripts for debugging delay accounting.
delayacct.py is designed to be used with current (e.g. 4.13.0) kernels.
delayacct_new.py is meant for a patched version of the kernel.

They each print the pointer of the delays struct that is updated when block I/O
ends. As an example:

 <idle>-0     [005] d.h. 68976.200976: 0x00000001: Updating stats at: ffff8881bb40e280

On nodes that have the bug, the running process (first column) directly
determines which delays struct is updated. On nodes that where the problem is
fixed, the running process has no effect on the updated pointer.